+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fix bug where aircraft cleared twice for ILS won't join glideslope [#667](https://github.com/openscope/openscope/issues/667)
 - Fix bug of aircraft descending via STAR to '0' altitude [#567](https://github.com/openscope/openscope/issues/567)
 - `sid` command no longer sets the aircraft's destination property [#787](https://github.com/openscope/openscope/issues/787)
+- Ensure aircraft future path is drawn only when directed by settings menu [#812](https://github.com/openscope/openscope/issues/812)